Split up test factories
I want to add more to these factories, so it'll be easier to have them
separate.
