+++ spec/javascript/_test_helpers/factories/id.ts
@@ -1,0 +1,8 @@
+let idCount = 0;
+
+function nextId() {
+  idCount += 1;
+  return idCount;
+}
+
+export {nextId};
