+++ spec/javascript/_test_helpers/factories/timeframe.ts
@@ -1,0 +1,12 @@
+function makeTimeframe(attrs: Partial<Timeframe>): Timeframe {
+  return {
+    name: 'inbox',
+    medianProductivity: 30,
+    minuteMax: Infinity,
+    currentTasks: [],
+    pendingTasks: [],
+    ...attrs,
+  };
+}
+
+export {makeTimeframe};
